# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Zak Ecosystem AI - CRM Omnicanale API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /b2b/competitor:
    get:
      operationId: listB2BCompetitor
      tags: [b2b]
      summary: List B2B competitor archive entries
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoria
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/B2BCompetitor"
    post:
      operationId: createB2BCompetitor
      tags: [b2b]
      summary: Create a B2B competitor archive entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/B2BCompetitorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B2BCompetitor"
  /b2b/competitor/{id}:
    get:
      operationId: getB2BCompetitor
      tags: [b2b]
      summary: Get a B2B competitor archive entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B2BCompetitor"
        "404":
          description: Competitor not found
    patch:
      operationId: updateB2BCompetitor
      tags: [b2b]
      summary: Update a B2B competitor archive entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/B2BCompetitorUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B2BCompetitor"
        "404":
          description: Competitor not found
    delete:
      operationId: deleteB2BCompetitor
      tags: [b2b]
      summary: Delete a B2B competitor archive entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /b2b/materiali:
    get:
      operationId: listB2BMateriali
      tags: [b2b]
      summary: List B2B competitor materials metadata
      parameters:
        - name: competitor_id
          in: query
          schema:
            type: string
        - name: stato
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/B2BMateriale"
    post:
      operationId: createB2BMateriale
      tags: [b2b]
      summary: Register competitor material metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/B2BMaterialeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B2BMateriale"
  /b2b/materiali/{id}:
    patch:
      operationId: updateB2BMateriale
      tags: [b2b]
      summary: Update competitor material metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/B2BMaterialeUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B2BMateriale"
        "404":
          description: Material not found
    delete:
      operationId: deleteB2BMateriale
      tags: [b2b]
      summary: Delete competitor material metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /b2b/template:
    get:
      operationId: listB2BTemplate
      tags: [b2b]
      summary: List B2B co-branding templates
      parameters:
        - name: target_tipo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/B2BTemplate"
    post:
      operationId: createB2BTemplate
      tags: [b2b]
      summary: Create a B2B co-branding template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/B2BTemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B2BTemplate"
  /b2b/template/{id}:
    patch:
      operationId: updateB2BTemplate
      tags: [b2b]
      summary: Update a B2B co-branding template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/B2BTemplateUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B2BTemplate"
        "404":
          description: Template not found
    delete:
      operationId: deleteB2BTemplate
      tags: [b2b]
      summary: Delete a B2B co-branding template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /b2b/analisi-competitor:
    post:
      operationId: analyzeB2BCompetitor
      tags: [b2b]
      summary: Generate a structured competitor analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/B2BCompetitorAnalysisInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B2BCompetitorAnalysisResult"
  /b2b/export:
    post:
      operationId: exportB2BPitch
      tags: [b2b]
      summary: Export a B2B pitch as PDF or presentation outline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/B2BExportInput"
      responses:
        "200":
          description: Export generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B2BExportResult"
...
  /calendar/google/sync:
    post:
      operationId: syncGoogleCalendar
      tags: [calendar]
      summary: Run Google Calendar synchronization
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleCalendarSyncInput"
      responses:
        "200":
          description: Synchronization result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoogleCalendarSyncResult"
...
components:
  schemas:
...
    B2BCompetitor:
      type: object
      required: [id, nome, categoria, data_creazione, data_aggiornamento]
      properties:
        id:
          type: string
        nome:
          type: string
        categoria:
          type: string
        citta:
          type: ["string", "null"]
        zona:
          type: ["string", "null"]
        target:
          type: ["string", "null"]
        prezzo_medio:
          type: ["integer", "null"]
        rating:
          type: ["integer", "null"]
        punti_forza:
          type: ["string", "null"]
        punti_deboli:
          type: ["string", "null"]
        sito:
          type: ["string", "null"]
        instagram:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        data_creazione:
          type: string
        data_aggiornamento:
          type: string
    B2BCompetitorInput:
      type: object
      required: [nome]
      properties:
        nome:
          type: string
        categoria:
          type: string
        citta:
          type: string
        zona:
          type: string
        target:
          type: string
        prezzo_medio:
          type: integer
        rating:
          type: integer
        punti_forza:
          type: string
        punti_deboli:
          type: string
        sito:
          type: string
        instagram:
          type: string
        note:
          type: string
    B2BCompetitorUpdate:
      type: object
      properties:
        nome:
          type: string
        categoria:
          type: string
        citta:
          type: string
        zona:
          type: string
        target:
          type: string
        prezzo_medio:
          type: integer
        rating:
          type: integer
        punti_forza:
          type: string
        punti_deboli:
          type: string
        sito:
          type: string
        instagram:
          type: string
        note:
          type: string
    B2BMateriale:
      type: object
      required: [id, nome_file, tipo_materiale, stato, data_creazione]
      properties:
        id:
          type: string
        competitor_id:
          type: ["string", "null"]
        competitor_nome:
          type: ["string", "null"]
        nome_file:
          type: string
        tipo_materiale:
          type: string
        url:
          type: ["string", "null"]
        stato:
          type: string
        note:
          type: ["string", "null"]
        data_creazione:
          type: string
    B2BMaterialeInput:
      type: object
      required: [nome_file]
      properties:
        competitor_id:
          type: string
        nome_file:
          type: string
        tipo_materiale:
          type: string
        url:
          type: string
        stato:
          type: string
        note:
          type: string
    B2BMaterialeUpdate:
      type: object
      properties:
        competitor_id:
          type: string
        nome_file:
          type: string
        tipo_materiale:
          type: string
        url:
          type: string
        stato:
          type: string
        note:
          type: string
    B2BTemplate:
      type: object
      required: [id, titolo, target_tipo, messaggio, utilizzi, data_creazione, data_aggiornamento]
      properties:
        id:
          type: string
        titolo:
          type: string
        target_tipo:
          type: string
        target_descrizione:
          type: ["string", "null"]
        messaggio:
          type: string
        vantaggi:
          type: ["string", "null"]
        cta:
          type: ["string", "null"]
        utilizzi:
          type: integer
        data_creazione:
          type: string
        data_aggiornamento:
          type: string
    B2BTemplateInput:
      type: object
      required: [titolo, messaggio]
      properties:
        titolo:
          type: string
        target_tipo:
          type: string
        target_descrizione:
          type: string
        messaggio:
          type: string
        vantaggi:
          type: string
        cta:
          type: string
        utilizzi:
          type: integer
    B2BTemplateUpdate:
      type: object
      properties:
        titolo:
          type: string
        target_tipo:
          type: string
        target_descrizione:
          type: string
        messaggio:
          type: string
        vantaggi:
          type: string
        cta:
          type: string
        utilizzi:
          type: integer
    B2BCompetitorAnalysisInput:
      type: object
      required: [prompt]
      properties:
        competitor_id:
          type: string
        prompt:
          type: string
        focus:
          type: string
          enum: [prezzo, proposta, debolezze, generale]
    B2BCompetitorAnalysisResult:
      type: object
      required: [titolo, sintesi, punti_forza, punti_deboli, opportunita, azioni_consigliate]
      properties:
        titolo:
          type: string
        sintesi:
          type: string
        punti_forza:
          type: array
          items:
            type: string
        punti_deboli:
          type: array
          items:
            type: string
        opportunita:
          type: array
          items:
            type: string
        azioni_consigliate:
          type: array
          items:
            type: string
        confidence:
          type: string
    B2BExportInput:
      type: object
      required: [titolo, target, formato]
      properties:
        titolo:
          type: string
        target:
          type: string
        messaggio:
          type: string
        budget:
          type: number
        formato:
          type: string
          enum: [pdf, presentazione]
    B2BExportResult:
      type: object
      required: [formato, titolo, contenuto, slides]
      properties:
        formato:
          type: string
          enum: [pdf, presentazione]
        titolo:
          type: string
        contenuto:
          type: string
        download_filename:
          type: string
        slides:
          type: array
          items:
            type: object
            required: [titolo, contenuto]
            properties:
              titolo:
                type: string
              contenuto:
                type: string
...
    GoogleCalendarSyncInput:
      type: object
      properties:
        direction:
          type: string
          enum: [zak_to_google, google_to_zak, bidirectional]
          default: bidirectional
        full_sync:
          type: boolean
          default: false
        days_ahead:
          type: integer
          default: 180
    GoogleCalendarSyncResult:
      type: object
      required: [configured, direction, pushed, pulled, skipped, conflicts, errors]
      properties:
        configured:
          type: boolean
        direction:
          type: string
        pushed:
          type: integer
        pulled:
          type: integer
        skipped:
          type: integer
        conflicts:
          type: integer
        errors:
          type: array
          items:
            type: string
        message:
          type: string
...